+++ packages/esp_board_manager/include/esp_board_periph.h
@@ -7,10 +7,7 @@
 
 #pragma once
 
-#include <stdbool.h>
-#include "driver/i2c.h"
-#include "driver/i2s_std.h"
-#include "driver/gpio.h"
+#include "esp_err.h"
 
 #ifdef __cplusplus
 extern "C" {